+++ Icon/Icon.js
@@ -350,6 +350,48 @@ const IconBase = kind({
  * exclamation
  * show
  * hide
+ * wifi15g
+ * wifi25g
+ * wifi35g
+ * wifi45g
+ * wifilock15g
+ * wifilock25g
+ * wifilock35g
+ * wifilock45g
+ * create
+ * background
+ * pen
+ * eraser
+ * mediaplayer
+ * location
+ * exit
+ * movies
+ * mycontents
+ * accessibility
+ * seemore
+ * trailer
+ * onnow
+ * tvguidefvp
+ * pcnotconnected
+ * fileppt
+ * textinput
+ * pagewidth
+ * index
+ * colorpicker
+ * ear
+ * hand
+ * remotecontrol
+ * channelscheduling
+ * smartfunction
+ * appscontents
+ * deviceconnect
+ * move
+ * voiced
+ * auracast
+ * jumpbackward10
+ * jumpforward10
+ * logout
+ * sketch
  * ```
  *
  * @name iconList