+++ tests/ui/specs/VirtualList/VirtualList-utils.js
@@ -15,15 +15,15 @@ async function hitTest (_selector) {
 	}, _selector);
 }
 
-async function expectFocusedItem (itemNum, comment = 'focused item') {
+async function expectFocusedItem (itemNum) {
 	const focusedId = await focusedElement();
-	expect(focusedId, comment).to.equal(`item${itemNum}`);
+	expect(focusedId).toBe(`item${itemNum}`);
 }
 
 async function expectNoFocusedItem () {
 	expect(await browser.execute(function () {
 		return document.activeElement === document.body;
-	})).to.be.true();
+	})).toBe(true);
 }
 
 async function waitUntilFocused (itemNum, comment = '') {
@@ -46,11 +46,11 @@ async function waitUntilVisible (itemNum) {
 }
 
 async function isScrolling () {
-	return await $('#scrolling').getText() === 'Scrolling';
+	return (await $('#scrolling').getText()) === 'Scrolling';
 }
 
 async function isNotScrolling () {
-	return await $('#scrolling').getText() === 'Not Scrolling';
+	return (await $('#scrolling').getText()) === 'Not Scrolling';
 }
 
 /**